Move all tests together to share helpers and fixtures
