+++ tests/__helpers__/rm.js
@@ -1,0 +1,4 @@
+const {promisify} = require('util')
+const rimraf = require('rimraf')
+
+module.exports = promisify(rimraf)
