+++ tests/util/archive.test.js
@@ -1,11 +1,8 @@
 const path = require('path')
-const {promisify} = require('util')
 const test = require('ava')
-const rimraf = require('rimraf')
-const {isArchive, unarchive} = require('../archive')
-const {createTempDirectory} = require('../tmpdir')
-
-const rm = promisify(rimraf)
+const {isArchive, unarchive} = require('../../lib/util/archive')
+const {createTempDirectory} = require('../../lib/util/tmpdir')
+const rm = require('../__helpers__/rm')
 
 test('should return false if there are no types specified', t => {
   t.is(isArchive([]), false)
@@ -36,29 +33,29 @@ test('should return true if there is at least one archive type', t => {
 
 test('should error when trying to unarchive a file that does not exist', async t => {
   await t.throws(
-    unarchive(path.resolve(__dirname, 'fixtures/doesnt-exist')),
+    unarchive(path.resolve(__dirname, '../__fixtures__/doesnt-exist')),
     'Could not extract archive'
   )
 })
 
 test('should error when trying to unarchive an invalid archive', async t => {
   await t.throws(
-    unarchive(path.resolve(__dirname, 'fixtures/non-archive.txt')),
+    unarchive(path.resolve(__dirname, '../__fixtures__/file.txt')),
     'Could not extract archive'
   )
 })
 
 test('should extract an archive and return the unarchived path', async t => {
-  const tmp = await unarchive(path.resolve(__dirname, 'fixtures/archive.zip'))
+  const tmp = await unarchive(path.resolve(__dirname, '../__fixtures__/file.zip'))
 
-  t.is(tmp, path.resolve(__dirname, 'fixtures/_unarchived'))
+  t.is(tmp, path.resolve(__dirname, '../__fixtures__/_unarchived'))
 
   return rm(tmp)
 })
 
 test('should allow extracting in a different path', async t => {
   const tmp = await createTempDirectory()
-  const ret = await unarchive(path.resolve(__dirname, 'fixtures/archive.zip'), tmp)
+  const ret = await unarchive(path.resolve(__dirname, '../__fixtures__/file.zip'), tmp)
 
   t.is(ret, path.join(tmp, '_unarchived'))
