+++ tests/util/tmpdir.test.js
@@ -1,11 +1,9 @@
 const os = require('os')
 const path = require('path')
-const {promisify} = require('util')
 const test = require('ava')
-const rimraf = require('rimraf')
-const tmpdir = require('../tmpdir')
+const tmpdir = require('../../lib/util/tmpdir')
+const rm = require('../__helpers__/rm')
 
-const rm = promisify(rimraf)
 const ostmpdir = os.tmpdir()
 
 test('should create a tmp directory prefixed with plunger_', async t => {
